Combine end router modules
move version 2 endpoint definitions to single module
move version 1 endpoint definitions to single module
