--- app/router/v1/__init__.py
@@ -1,4 +1,0 @@
-"""app.router.v1"""
-from fastapi import APIRouter
-
-V1 = APIRouter()
