--- app/router/v1/deaths.py
@@ -1,11 +1,0 @@
-"""app.router.v1.deaths.py"""
-from ...services.location.jhu import get_category
-from . import V1
-
-
-@V1.get("/deaths")
-async def deaths():
-    """Total deaths."""
-    deaths_data = await get_category("deaths")
-
-    return deaths_data
