--- app/router/v2/__init__.py
@@ -1,4 +1,0 @@
-"""app.router.v2"""
-from fastapi import APIRouter
-
-V2 = APIRouter()
