+++ app/routers/__init__.py
@@ -1,0 +1,3 @@
+"""app.routers"""
+from .v1 import V1
+from .v2 import V2
